created symbolic links for header files to be used in spm
