+++ AnalyticsMVCID/include/AnalyticsMVCID.h
@@ -1,20 +1,1 @@
-//
-//  AnalyticsMVCID.h
-//  AnalyticsMVCID
-//
-//  Created by Hector Rodriguez on 2/4/20.
-//  Copyright © 2020 Segment. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for AnalyticsMVCID.
-FOUNDATION_EXPORT double AnalyticsMVCIDVersionNumber;
-
-//! Project version string for AnalyticsMVCID.
-FOUNDATION_EXPORT const unsigned char AnalyticsMVCIDVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <AnalyticsMVCID/PublicHeader.h>
-
-#import "SEGMCVIDTracker.h"
-
+../AnalyticsMVCID.h
